Changes: * rename navbars * refactors structure director Navbars
