+++ src/components/Navbars/UserNavbar/UserNavbar.js
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./OwnerNavbar.module.scss";
+import styles from "./UserNavbar.module.scss";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faBars } from "@fortawesome/free-solid-svg-icons";
 import UserData from "../../UserData/UserData";
